added refresh token generation when user registered

+++ wellony-api/src/main/java/git/whitechoke/wellony/security/JwtUtils.java
@@ -29,7 +29,7 @@ public class JwtUtils {
     @Value("${jwt.access-expiry-ms}")
     private int accessExpiryMs;
     @Value("${jwt.refresh-expiry-ms}")
-    private int refreshExpiryMs;
+    private long refreshExpiryMs;
 
     private final RefreshTokenRepository refreshTokenRepository;
